planner: move PhysicalApply related logic into physical op dir
