+++ pkg/planner/core/physical_plans.go
@@ -16,7 +16,6 @@ package core
 
 import (
 	"fmt"
-	"slices"
 	"strconv"
 	"strings"
 	"unsafe"
@@ -69,7 +68,7 @@ var (
 	_ base.PhysicalPlan = &PhysicalIndexMergeReader{}
 	_ base.PhysicalPlan = &physicalop.PhysicalHashAgg{}
 	_ base.PhysicalPlan = &PhysicalStreamAgg{}
-	_ base.PhysicalPlan = &PhysicalApply{}
+	_ base.PhysicalPlan = &physicalop.PhysicalApply{}
 	_ base.PhysicalPlan = &physicalop.PhysicalIndexJoin{}
 	_ base.PhysicalPlan = &physicalop.PhysicalHashJoin{}
 	_ base.PhysicalPlan = &physicalop.PhysicalMergeJoin{}
@@ -835,60 +834,6 @@ func expandVirtualColumn(schema *expression.Schema, copyColumn []*model.ColumnIn
 		}
 	}
 	return copyColumn
-}
-
-// PhysicalApply represents apply plan, only used for subquery.
-type PhysicalApply struct {
-	physicalop.PhysicalHashJoin
-
-	CanUseCache bool
-	Concurrency int
-	OuterSchema []*expression.CorrelatedColumn
-}
-
-// PhysicalJoinImplement has an extra bool return value compared with PhysicalJoin interface.
-// This will override BasePhysicalJoin.PhysicalJoinImplement() and make PhysicalApply not an implementation of
-// base.PhysicalJoin interface.
-func (*PhysicalApply) PhysicalJoinImplement() bool { return false }
-
-// Clone implements op.PhysicalPlan interface.
-func (la *PhysicalApply) Clone(newCtx base.PlanContext) (base.PhysicalPlan, error) {
-	cloned := new(PhysicalApply)
-	cloned.SetSCtx(newCtx)
-	base, err := la.PhysicalHashJoin.Clone(newCtx)
-	if err != nil {
-		return nil, err
-	}
-	hj := base.(*physicalop.PhysicalHashJoin)
-	cloned.PhysicalHashJoin = *hj
-	cloned.CanUseCache = la.CanUseCache
-	cloned.Concurrency = la.Concurrency
-	for _, col := range la.OuterSchema {
-		cloned.OuterSchema = append(cloned.OuterSchema, col.Clone().(*expression.CorrelatedColumn))
-	}
-	return cloned, nil
-}
-
-// ExtractCorrelatedCols implements op.PhysicalPlan interface.
-func (la *PhysicalApply) ExtractCorrelatedCols() []*expression.CorrelatedColumn {
-	corCols := la.PhysicalHashJoin.ExtractCorrelatedCols()
-	return slices.DeleteFunc(corCols, func(col *expression.CorrelatedColumn) bool {
-		return la.Children()[0].Schema().Contains(&col.Column)
-	})
-}
-
-// MemoryUsage return the memory usage of PhysicalApply
-func (la *PhysicalApply) MemoryUsage() (sum int64) {
-	if la == nil {
-		return
-	}
-
-	sum = la.PhysicalHashJoin.MemoryUsage() + size.SizeOfBool + size.SizeOfBool + size.SizeOfSlice +
-		int64(cap(la.OuterSchema))*size.SizeOfPointer
-	for _, corrCol := range la.OuterSchema {
-		sum += corrCol.MemoryUsage()
-	}
-	return
 }
 
 // PhysicalJoin provides some common methods for join operators.
