+++ pkg/planner/core/resolve_indices.go
@@ -398,49 +398,6 @@ func resolveIndices4PhysicalLimit(pp base.PhysicalPlan) (err error) {
 }
 
 // ResolveIndices implements Plan interface.
-func (p *PhysicalApply) ResolveIndices() (err error) {
-	err = p.PhysicalHashJoin.ResolveIndices()
-	if err != nil {
-		return err
-	}
-	// p.OuterSchema may have duplicated CorrelatedColumns,
-	// we deduplicate it here.
-	dedupCols := make(map[int64]*expression.CorrelatedColumn, len(p.OuterSchema))
-	for _, col := range p.OuterSchema {
-		dedupCols[col.UniqueID] = col
-	}
-	p.OuterSchema = make([]*expression.CorrelatedColumn, 0, len(dedupCols))
-	for _, col := range dedupCols {
-		newCol, err := col.Column.ResolveIndices(p.Children()[0].Schema())
-		if err != nil {
-			return err
-		}
-		col.Column = *newCol.(*expression.Column)
-		p.OuterSchema = append(p.OuterSchema, col)
-	}
-	// Resolve index for equal conditions again, because apply is different from
-	// hash join on the fact that equal conditions are evaluated against the join result,
-	// so columns from equal conditions come from merged schema of children, instead of
-	// single child's schema.
-	joinedSchema := expression.MergeSchema(p.Children()[0].Schema(), p.Children()[1].Schema())
-	for i, cond := range p.PhysicalHashJoin.EqualConditions {
-		newSf, err := cond.ResolveIndices(joinedSchema)
-		if err != nil {
-			return err
-		}
-		p.PhysicalHashJoin.EqualConditions[i] = newSf.(*expression.ScalarFunction)
-	}
-	for i, cond := range p.PhysicalHashJoin.NAEqualConditions {
-		newSf, err := cond.ResolveIndices(joinedSchema)
-		if err != nil {
-			return err
-		}
-		p.PhysicalHashJoin.NAEqualConditions[i] = newSf.(*expression.ScalarFunction)
-	}
-	return
-}
-
-// ResolveIndices implements Plan interface.
 func (p *Update) ResolveIndices() (err error) {
 	err = p.SimpleSchemaProducer.ResolveIndices()
 	if err != nil {
